Add files via upload
Program 1: Name: V1chattingRobot
Write a simple chatting-robot program. Make the user believe that he/she
is talking to an intelligent machine. In V1 the computer leads the conversation
by asking the user questions and answering them in an intelligent manner.
Use the time module to pause between questions and answers.